remove unused annotations, make one common configuration file, write some rabbit config to properties
