+++ src/main/java/com/instahipsta/harCRUD/Application.java
@@ -1,11 +1,11 @@
 package com.instahipsta.harCRUD;
 
-import com.instahipsta.harCRUD.config.RabbitConfiguration;
+import com.instahipsta.harCRUD.config.AppConfiguration;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.context.annotation.Import;
 
-@Import(RabbitConfiguration.class)
+@Import(AppConfiguration.class)
 @SpringBootApplication
 public class Application {
     public static void main(String[] args) {
