+++ src/main/java/com/instahipsta/harCRUD/config/AppConfiguration.java
@@ -1,0 +1,56 @@
+package com.instahipsta.harCRUD.config;
+
+import lombok.AllArgsConstructor;
+import lombok.RequiredArgsConstructor;
+import org.modelmapper.ModelMapper;
+import org.springframework.amqp.core.Binding;
+import org.springframework.amqp.core.BindingBuilder;
+import org.springframework.amqp.core.DirectExchange;
+import org.springframework.amqp.core.Queue;
+import org.springframework.amqp.rabbit.annotation.EnableRabbit;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+import static org.modelmapper.config.Configuration.AccessLevel.PRIVATE;
+import static org.modelmapper.convention.MatchingStrategies.STRICT;
+
+@Configuration
+@EnableRabbit
+//@RequiredArgsConstructor
+public class AppConfiguration {
+
+    @Value("${spring.rabbitmq.template.default-receive-queue}")
+    private String queueName;
+    @Value("${spring.rabbitmq.template.exchange}")
+    private String exchangeName;
+    @Value("${spring.rabbitmq.template.routing-key}")
+    private String routingKeyName;
+
+    @Bean
+    public ModelMapper mapper() {
+        ModelMapper mapper = new ModelMapper();
+        mapper.getConfiguration()
+                .setMatchingStrategy(STRICT)
+                .setFieldMatchingEnabled(true)
+                .setSkipNullEnabled(true)
+                .setFieldAccessLevel(PRIVATE);
+        return mapper;
+    }
+
+    @Bean
+    public Queue harQueue() {
+        return new Queue(queueName);
+    }
+
+    @Bean
+    public DirectExchange directExchange() {
+        return new DirectExchange(exchangeName);
+    }
+
+    @Bean
+    public Binding harBinding() {
+        return BindingBuilder.bind(harQueue()).to(directExchange()).with(routingKeyName);
+    }
+
+}
