--- src/main/java/com/instahipsta/harCRUD/config/ModelMapperConfiguration.java
@@ -1,24 +1,0 @@
-package com.instahipsta.harCRUD.config;
-
-import org.modelmapper.ModelMapper;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-import static org.modelmapper.config.Configuration.AccessLevel.PRIVATE;
-import static org.modelmapper.convention.MatchingStrategies.STRICT;
-
-@Configuration
-public class ModelMapperConfiguration {
-
-    @Bean
-    public ModelMapper mapper() {
-        ModelMapper mapper = new ModelMapper();
-        mapper.getConfiguration()
-                .setMatchingStrategy(STRICT)
-                .setFieldMatchingEnabled(true)
-                .setSkipNullEnabled(true)
-                .setFieldAccessLevel(PRIVATE);
-        return mapper;
-    }
-
-}
