+++ src/main/java/com/instahipsta/harCRUD/service/impl/FileServiceImpl.java
@@ -13,7 +13,6 @@ import java.util.UUID;
 
 @Service
 @Slf4j
-@RequiredArgsConstructor
 public class FileServiceImpl implements FileService {
 
     @Value("${file.downloads}")
