+++ src/test/resourses/application.yml
@@ -20,6 +20,10 @@ spring:
     host: 127.0.0.1
     username: root
     password: root
+    template:
+      exchange: harExchange
+      default-receive-queue: harQueue
+      routing-key: har
   liquibase:
     enabled: true
     change-log: classpath*:db.changelog/db.changelog-master-dev.yaml
